# Verdi: Load config / profile lazily to speed up tab-completion